move clickhouse cluster resource into common location
