+++ dags/definitions.py
@@ -12,8 +12,9 @@ from dagster import fs_io_manager
 from django.conf import settings
 
 from . import ch_examples, deletes, orm_examples
+from .common import ClickhouseClusterResource
 from .materialized_columns import materialize_column
-from .person_overrides import ClickhouseClusterResource, squash_person_overrides
+from .person_overrides import squash_person_overrides
 
 all_assets = load_assets_from_modules([ch_examples, orm_examples])
