+++ dags/deletes.py
@@ -9,9 +9,7 @@ from dagster import (
     OpExecutionContext,
     Config,
     MetadataValue,
-    InitResourceContext,
     ResourceParam,
-    ConfigurableResource,
 )
 from django.conf import settings
 from functools import partial
@@ -22,27 +20,9 @@ from posthog.clickhouse.cluster import (
     Mutation,
     MutationRunner,
     NodeRole,
-    get_cluster,
 )
 from posthog.models.event.sql import EVENTS_DATA_TABLE
 from posthog.models.async_deletion import AsyncDeletion, DeletionType
-
-
-class ClickhouseClusterResource(ConfigurableResource):
-    """
-    The ClickHouse cluster used to run the job.
-    """
-
-    client_settings: dict[str, str] = {
-        "max_execution_time": "0",
-        "max_memory_usage": "0",
-        "receive_timeout": f"{24 * 60 * 60}",  # wait 24 hours for a response from CH
-        "mutations_sync": "0",
-        "lightweight_deletes_sync": "0",
-    }
-
-    def create_resource(self, context: InitResourceContext) -> ClickhouseCluster:
-        return get_cluster(context.log, client_settings=self.client_settings)
 
 
 class DeleteConfig(Config):
